+++ examples/website/icon/icon-cluster-layer.js
@@ -28,7 +28,7 @@ export default class IconClusterLayer extends CompositeLayer {
     const rebuildIndex = changeFlags.dataChanged || props.sizeScale !== oldProps.sizeScale;
 
     if (rebuildIndex) {
-      const index = new Supercluster({maxZoom: 16, radius: props.sizeScale});
+      const index = new Supercluster({maxZoom: 16, radius: props.sizeScale * Math.sqrt(2)});
       index.load(
         props.data.map(d => ({
           geometry: {coordinates: props.getPosition(d)},