+++ modules/carto/src/api/index.js
@@ -1,3 +1,3 @@
-export {FORMATS, MAP_TYPES, PROVIDERS, API_VERSIONS} from './maps-api-common';
+export {FORMATS, MAP_TYPES, API_VERSIONS} from './maps-api-common';
 export {getMapCarto, CONNECTIONS} from './maps-client';
-export {getMapCartoCloudNative} from './maps-cloud-native-client';
+export {getMapCartoCloudNative, getMapMetadata} from './maps-cloud-native-client';