+++ modules/carto/src/layers/carto-layer.js
@@ -2,8 +2,8 @@ import {CompositeLayer, log} from '@deck.gl/core';
 import {MVTLayer} from '@deck.gl/geo-layers';
 import {GeoJsonLayer} from '@deck.gl/layers';
 import {getMapCartoCloudNative, getMapCarto, FORMATS, API_VERSIONS} from '../api';
-import {MAP_TYPES, PROVIDERS} from '../api/maps-api-common';
-import {getConfig} from '../config';
+import {MAP_TYPES} from '../api/maps-api-common';
+import {getDefaultCredentials} from '../config';
 
 const defaultProps = {
   // (String, required): data resource to load. table name, sql query or tileset name.
@@ -16,14 +16,11 @@ const defaultProps = {
   /*********************/
   /* API v3 PARAMETERS */
   /**********************/
-
-  // (String {bigquery, snowflake,redshift, postgres}, required)
-  provider: null,
   // (String, required): connection name at CARTO platform
   connection: null,
 
-  // override carto config for the layer, set to null to read from default
-  config: null,
+  // override carto credentials for the layer, set to null to read from default
+  credentials: null,
   // sublayer used to render. Any deck.gl layer or null to autodetect
   renderSubLayer: null,
   // (String {geojson, json, tileset}, optional). Desired data format. By default, it's guessed automaticaly
@@ -43,9 +40,9 @@ export default class CartoLayer extends CompositeLayer {
   }
 
   _checkProps(props) {
-    const {provider, type, format, config} = this.props;
-    const localConfig = {...getConfig(), ...config};
-    const {apiVersion} = localConfig;
+    const { type, format, credentials, connection } = this.props;
+    const localCreds = {...getDefaultCredentials(), ...credentials};
+    const {apiVersion} = localCreds;
 
     log.assert(
       Object.values(API_VERSIONS).includes(apiVersion),
@@ -55,20 +52,19 @@ export default class CartoLayer extends CompositeLayer {
     );
 
     if (apiVersion === API_VERSIONS.V1 || apiVersion === API_VERSIONS.V2) {
-      log.assert(!provider, `Provider not suport for apiVersion ${apiVersion}`);
-      log.assert(!format, `Format not suport for apiVersion ${apiVersion}`);
+      log.assert(!format, `Format not supported for apiVersion ${apiVersion}`);
       log.assert(
         type !== MAP_TYPES.TABLE,
-        `Use type ${MAP_TYPES.SQL} or ${MAP_TYPES.TILESET} for apiVersion ${apiVersion}`
+        `Use type ${MAP_TYPES.QUERY} or ${MAP_TYPES.TILESET} for apiVersion ${apiVersion}`
+      )
+      log.assert(
+        !connection,
+        `Connection prop is not supported for apiVersion ${apiVersion}`
       );
     } else if (apiVersion === API_VERSIONS.V3) {
       log.assert(
         Object.values(MAP_TYPES).includes(type),
         `Invalid type ${type}. Possible values are ${Object.values(MAP_TYPES).toString()}`
-      );
-      log.assert(
-        Object.values(PROVIDERS).includes(provider),
-        `Invalid provider ${provider}. Possible values are ${Object.values(PROVIDERS).toString()}`
       );
       log.assert(
         !format || Object.values(FORMATS).includes(format),
@@ -81,10 +77,9 @@ export default class CartoLayer extends CompositeLayer {
     this._checkProps(props);
     const shouldUpdateData =
       changeFlags.dataChanged ||
-      props.provider !== oldProps.provider ||
       props.connection !== oldProps.connection ||
       props.type !== oldProps.type ||
-      JSON.stringify(props.config) !== JSON.stringify(oldProps.config);
+      JSON.stringify(props.credentials) !== JSON.stringify(oldProps.credentials);
 
     if (shouldUpdateData) {
       this.setState({data: null, renderSubLayer: null});
@@ -94,8 +89,8 @@ export default class CartoLayer extends CompositeLayer {
 
   async _updateData() {
     try {
-      const {provider, type, data: source, connection, config, format} = this.props;
-      const localConfig = {...getConfig(), ...config};
+      const { type, data: source, connection, credentials, format} = this.props;
+      const localConfig = {...getDefaultCredentials(), ...credentials};
       const {apiVersion} = localConfig;
 
       let data;
@@ -103,15 +98,14 @@ export default class CartoLayer extends CompositeLayer {
 
       if (apiVersion === API_VERSIONS.V3) {
         [data, mapFormat] = await getMapCartoCloudNative({
-          provider,
           type,
           source,
           connection,
-          config,
+          credentials,
           format
         });
       } else if (apiVersion === API_VERSIONS.V1 || apiVersion === API_VERSIONS.V2) {
-        [data, mapFormat] = await getMapCarto({type, source, config});
+        [data, mapFormat] = await getMapCarto({type, source, credentials});
       } else {
         log.assert(
           `Unknow apiVersion ${apiVersion}. Possible values are ${Object.values(