+++ modules/core/src/lib/deck-picker.js
@@ -44,16 +44,11 @@ export default class DeckPicker {
       layerId: null,
       info: null
     };
-    this._onError = null;
   }
 
   setProps(props) {
     if ('layerFilter' in props) {
       this.layerFilter = props.layerFilter;
-    }
-
-    if ('onError' in props) {
-      this._onError = props.onError;
     }
 
     if ('_pickable' in props) {
@@ -379,7 +374,6 @@ export default class DeckPicker {
     this.pickLayersPass.render({
       layers,
       layerFilter: this.layerFilter,
-      onError: this._onError,
       views,
       viewports,
       onViewportActive,